add:js-snack7

+++ js-snack/snack-06.js
@@ -9,7 +9,7 @@ Fate generare un numero random da 0 a 10 al computer,
  let numeroCasuale = Math.round(Math.random() * 10 )
  console.log(numeroCasuale)
 
- let numeroUtente = parseFloat(prompt('inserisci un numero da 1 a 10'))
+ let numeroUtente = parseInt(prompt('inserisci un numero da 1 a 10'))
 
  if(numeroUtente === numeroCasuale ){
     console.log('Hai vinto!')
